rename the mixin plugin class and clean up the compat mixin checks
