+++ gradle/forge.versions.toml
@@ -1,5 +1,5 @@
 [versions]
-ldlib = "1.0.39"
+ldlib = "1.0.40.a"
 registrate = "MC1.20-1.3.11"
 configuration = "2.2.0"
 mixinExtras = "0.5.0-rc.3"
